Fix calling {p,r,}partial with no args
The no args test actually passed null to the partial,
which started to fail as PHP got stricter in 8.0.

Let’s actually make partial function accept no arguments
to get rid of the edge case and then we will be able to test it properly.

Let’s also return to the original time function for consistency:
https://github.com/ihor/NSPL/commit/cd9e331494c1f90e016f8c1ded6eff8f28e68f50